test(windows): split native retry matrix within per-file runtime budget
